Changed duplicated currency and count formatting to shared helpers
no ref

Settings carried a byte-identical copy of the framework's getSymbol plus
the currency list and decimal/minimum-amount helpers; those now live in
admin-x-framework's currency util, exported from the root barrel, with
settings importing them from there. The duplicate centsToDollars copies
in post analytics growth (a local helper file and an inline closure) now
point at the Shade export, and raw toLocaleString count formatting in the
members bulk modals, import complete step, post success modal, and stats
filter goes through Shade's formatNumber like the rest of Admin.
