--- apps/admin/src/posts/analytics/growth/growth-helpers.ts
@@ -1,3 +1,0 @@
-export const centsToDollars = (value : number) => {
-    return Math.round(value / 100);
-};
