+++ apps/admin/src/posts/analytics/overview/overview.tsx
@@ -11,8 +11,7 @@ import {type KpiDataItem} from '@/posts/analytics/utils/kpi-helpers';
 import {LucideIcon, formatNumber} from '@tryghost/shade/utils';
 import {type Post, usePostAnalytics} from '@/posts/analytics/providers/post-analytics-context';
 import {STATS_RANGES} from '@/shared/analytics/constants';
-import {centsToDollars} from '@/posts/analytics/growth/growth-helpers';
-import {formatQueryDate, getRangeDates, getRangeForStartDate} from '@tryghost/shade/app';
+import {centsToDollars, formatQueryDate, getRangeDates, getRangeForStartDate} from '@tryghost/shade/app';
 import {getEffectiveChartRange, sanitizeChartData} from '@/shared/analytics/chart-helpers';
 import {hasBeenEmailed, isPublishedOnly, useNavigate, useTinybirdQuery} from '@tryghost/admin-x-framework';
 import {useActiveGiftLink} from '@tryghost/admin-x-framework/api/gift-links';
